refactor connect method
