+++ modules/client/src/index.ts
@@ -1,4 +1,5 @@
-import { connect, ConnextClient } from "./connext";
+import { connect } from "./connect";
+import { ConnextClient } from "./connext";
 import * as utils from "./lib";
 
 export { ConnextClient, connect, utils };
